Add randomized instance generator deps
Here's a concise commit message that captures the changes:

``` Add rand/rand_distr deps for instance generation

Required for randomly generating synthetic berth allocation instances
with: - Poisson process arrivals - Uniform length sampling - Normal
processing times ```

The changes add dependencies needed for implementing random instance
generation with statistical distributions. I kept the message focused on
the direct changes (adding dependencies) and noted their purpose without
repeating the full diff.
