+++ crates/dock-alloc-model/Cargo.toml
@@ -27,3 +27,5 @@ edition = "2024"
 [dependencies]
 dock_alloc_core = { path = "../dock-alloc-core" }
 num-traits = "0.2.19"
+rand = "0.9.2"
+rand_distr = "0.5.1"
